+++ packages/address-picker/src/index.style.tsx
@@ -1,11 +1,9 @@
 import styled from 'styled-components';
 import { Info } from '@lendi-ui/icon';
-import { container } from '@lendi-ui/container';
 import { pt } from '@lendi-ui/spacing';
 
 export const Wrapper = styled.section`
-  ${container()}
-  max-width: 385px;
+  width: 100%;
 `;
 
 export const Alert = styled.div`